added interactive loop

+++ lmntfy.py
@@ -8,6 +8,7 @@ data_folder = './data'
 docs_folder = data_folder + '/docs'
 database_path_prefix = data_folder + '/database'
 rebuild_database = False
+use_test_questions = False
 
 llm = lmntfy.models.llm.GPT35()
 embedder = lmntfy.models.embedding.OpenAIEmbedding()
@@ -36,8 +37,14 @@ else:
     print("Loaded database from file.")
 
     question_answerer = lmntfy.QuestionAnswerer(llm, embedder, database)
-    test_questions = ["What is NERSC?", "How can I connect to Perlmutter?", "Where do I find gcc?", "How do I kill all of my jobs?", "How can I run a job on GPU?"]
-    for question in test_questions:
-        print(f"\n> {question}\n")
-        answer = question_answerer.get_answer(question, verbose=False)
-        print(f"\n{answer}\n")
+    if use_test_questions:
+        test_questions = ["What is NERSC?", "How can I connect to Perlmutter?", "Where do I find gcc?", "How do I kill all of my jobs?", "How can I run a job on GPU?"]
+        for question in test_questions:
+            print(f"\n> {question}\n")
+            answer = question_answerer.get_answer(question, verbose=False)
+            print(f"\n{answer}\n")
+    else:
+        while True:
+            question = input("\n> ")
+            answer = question_answerer.get_answer(question, verbose=False)
+            print(f"\n{answer}")
